Cleaned up - basic searching functionality, nothing fancy.
